Separate player detection logic from field of view logic. Ensure enemies only fire their weapons if they're looking in the player's general direction, per the field of view.
