+++ Assets/Scripts/Entity/Enemy/DetectPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b453d836d3397de41ba53dc5d59346c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
