+++ Assets/Scripts/Entity/Enemy/MovementBehaviors/PursuePlayer.cs
@@ -8,24 +8,24 @@ using UnityEngine;
 [RequireComponent(typeof(EnemyMovement))]
 public class PursuePlayer : MonoBehaviour, IOnDeathController
 {
-    private EnemyMovement em;
+    private EnemyMovement movement;
     [HideInInspector] public Transform playerTransform;
 
     public void OnDeath()
     {
         enabled = false;
-        em.Stop();
+        movement.Stop();
     }
 
     void Awake()
     {
-        em = GetComponent<EnemyMovement>();
+        movement = GetComponent<EnemyMovement>();
         playerTransform = GameObject.FindGameObjectWithTag("Player").transform;
     }
 
     // Update is called once per frame
     void Update()
     {
-        em.MoveTo(playerTransform.position);
+        movement.MoveTo(playerTransform.position);
     }
 }
